Multi-task, multi-dtype, naming updates...
This introduces multi-task. Now a sentence-transformers model is loaded once and supports all the text-embedding tasks.

Each task has a single/plural version.  This accomodates a desire to have a single batchable text/source_sentence/query which is common in caikit-nlp.
The plural version allows texts/source_sentences/queries so a batch-like interface is exposed. In particular the sentence-similarity and reranks "Tasks" take a list of source_sentences/queries but each is processed against the same list of docs -- not a list of docs per query!

Embedding vectors were expanded for future use of float64 and float16.

Naming and parameters have been update to latest -- still under review and subject to change.

Signed-off-by: markstur <mark.sturdevant@ibm.com>